EasyR1/requirements_b200.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-doc==0.0.3
antlr4-python3-runtime==4.9.3
anyio==4.11.0
astor==0.8.1
attrs==25.4.0
av==16.0.1
blake3==1.0.8
cachetools==6.2.1
cbor2==5.7.1
cfgv==3.4.0
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.8
compressed-tensors==0.11.0
cupy-cuda12x==13.6.0
datasets==4.3.0
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
fastapi==0.120.4
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
filelock==3.20.0
flashinfer-python==0.3.1
frozenlist==1.8.0
fsspec==2025.9.0
gguf==0.17.1
gitdb==4.0.12
gitpython==3.1.45
google-api-core==2.28.1
google-auth==2.42.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
identify==2.6.15
importlib-metadata==8.7.0
iniconfig==2.3.0
interegular==0.3.3
jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.2.2
liger-kernel==0.6.3
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
markdown==3.9
markupsafe==3.0.3
mathruler==0.1.0
mistral-common==1.8.5
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.15.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.6.1
openai-harmony==0.0.4
opencensus==0.11.4
opencensus-context==0.1.3
opencv-fixer==0.2.5
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-exporter-prometheus==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optree==0.17.0
outlines-core==0.2.11
pandas==2.3.3
partial-json-parser==0.2.1.1.post6
peft==0.17.1
pillow==12.0.0
pre-commit==4.3.0
prometheus-client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.2
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pybind11==3.0.1
pycountry==24.6.1
pydantic-extra-types==2.10.6
pylatexenc==2.10
pynvml==13.0.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
pyvers==0.1.0
pyyaml==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
ray==2.51.1
referencing==0.37.0
regex==2025.10.23
rich-toolkit==0.15.1
rignore==0.7.3
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.3
safetensors==0.6.2
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.43.0
setproctitle==1.3.7
setuptools==79.0.1
six==1.17.0
smart-open==7.4.2
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.10.0
tiktoken==0.12.0
tokenizers==0.22.1
torchdata==0.11.0
transformers==4.57.1
triton==3.4.0
tzdata==2025.2
uv==0.9.7
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
vllm==0.11.0
wandb==0.22.3
watchfiles==1.1.1
websockets==15.0.1
werkzeug==3.1.3
wrapt==2.0.0
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0

